Adds logic to prevent redundant swift downloads

+++ osx/install_swift_binaries.sh
@@ -26,14 +26,34 @@ set -e
 
 # Install OS X system level dependencies
 brew update > /dev/null
-#brew install curl
-brew install wget > /dev/null || brew outdated wget > /dev/null || brew upgrade wget > /dev/null
 
-# Install Swift binaries
-# See http://apple.stackexchange.com/questions/72226/installing-pkg-with-terminal
-# TODO: Since Xcode now includes the Swift compiler and Xcode is included in the macOS image provided
-# by Travis CI, we could add logic here that checks whether the Swift binaries are already available
-wget https://swift.org/builds/$SNAPSHOT_TYPE/xcode/$SWIFT_SNAPSHOT/$SWIFT_SNAPSHOT-osx.pkg
-sudo installer -pkg $SWIFT_SNAPSHOT-osx.pkg -target /
-export PATH=/Library/Developer/Toolchains/swift-latest.xctoolchain/usr/bin:"${PATH}"
-rm $SWIFT_SNAPSHOT-osx.pkg
+XCODE=$(/usr/bin/xcodebuild -version)
+XCODE_VERSION=$(echo ${XCODE} | sed -n 1p | cut -d' ' -f2)
+SWIFT_VERSION=$(cut -d '-' -f 2- <<< "swift-4.0.3-RELEASE" | tr -d '[:space:]')
+
+# Travis only supports 9.2, 9.1, 9.0, 8.3.3, 8.3, 7.3, and 6.4
+([ "${XCODE_VERSION}" == "9.2" ] && [ ${SWIFT_VERSION} == "4.0.3-RELEASE" ]) && alreadyDownloaded=1
+([ "${XCODE_VERSION}" == "9.1" ] && [ ${SWIFT_VERSION} == "4.0.2-RELEASE" ]) && alreadyDownloaded=1
+([ "${XCODE_VERSION}" == "9.0" ] && [ ${SWIFT_VERSION} == "4.0-RELEASE" ]) && alreadyDownloaded=1
+([ "${XCODE_VERSION}" == "8.3.3" ] && [ ${SWIFT_VERSION} == "3.1.1-RELEASE" ]) && alreadyDownloaded=1
+([ "${XCODE_VERSION}" == "8.3.2" ] && [ ${SWIFT_VERSION} == "3.1.1-RELEASE" ]) && alreadyDownloaded=1
+([ "${XCODE_VERSION}" == "8.3" ] && [ ${SWIFT_VERSION} == "3.1-RELEASE" ]) && alreadyDownloaded=1
+([ "${XCODE_VERSION}" == "8.2" ] && [ ${SWIFT_VERSION} == "3.0.2-RELEASE" ]) && alreadyDownloaded=1
+([ "${XCODE_VERSION}" == "8.1" ] && [ ${SWIFT_VERSION} == "3.0.1-RELEASE" ]) && alreadyDownloaded=1
+([ "${XCODE_VERSION}" == "8.0" ] && [ ${SWIFT_VERSION} == "3.0-RELEASE" ]) && alreadyDownloaded=1
+([ "${XCODE_VERSION}" == "7.3.1" ] && [ ${SWIFT_VERSION} == "2.2.1-RELEASE" ]) && alreadyDownloaded=1
+([ "${XCODE_VERSION}" == "7.3" ] && [ ${SWIFT_VERSION} == "2.2-RELEASE" ]) && alreadyDownloaded=1
+
+if [ ${alreadyDownloaded} ] ; then
+  echo "Swift ${SWIFT_SNAPSHOT} compatible with Xcode ${XCODE_VERSION}"
+else
+  # Install curl
+  brew install wget > /dev/null || brew outdated wget > /dev/null || brew upgrade wget > /dev/null
+
+  # Install Swift binaries
+  # See http://apple.stackexchange.com/questions/72226/installing-pkg-with-terminal
+  wget https://swift.org/builds/$SNAPSHOT_TYPE/xcode/$SWIFT_SNAPSHOT/$SWIFT_SNAPSHOT-osx.pkg
+  sudo installer -pkg $SWIFT_SNAPSHOT-osx.pkg -target /
+  export PATH=/Library/Developer/Toolchains/swift-latest.xctoolchain/usr/bin:"${PATH}"
+  rm $SWIFT_SNAPSHOT-osx.pkg
+fi
